ADDON-21053:added cluster_name functional test (#83)
