+++ .circleci/sck_values.yml
@@ -10,6 +10,7 @@ global:
 splunk-kubernetes-logging:
   kubernetes:
     insecureSSL: true
+    clusterName: "circleci-k8s-cluster"
   splunk:
     hec:
       indexName: circleci_events
