+++ test/conftest.py
@@ -1,5 +1,5 @@
 """
-Copyright 2018 Splunk, Inc..
+Copyright 2018-2019 Splunk, Inc..
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
